requirements.txt
# =============================================================================
# Whisper Vietnamese Fine-tuning - Dependencies
# =============================================================================

# --- Core ML ---
torch>=2.1.0
torchaudio>=2.1.0

# --- Hugging Face ---
transformers>=4.40.0
datasets>=2.18.0
peft>=0.10.0
accelerate>=0.28.0

# --- PyTorch Lightning ---
pytorch-lightning>=2.2.0
rich>=13.0.0

# --- Audio Processing ---
librosa>=0.10.0
soundfile>=0.12.0

# --- Evaluation ---
jiwer>=3.0.0

# --- Data ---
pandas>=2.0.0
pyarrow>=15.0.0
numpy>=1.24.0

# --- Configuration ---
pyyaml>=6.0

# --- Demo ---
gradio>=4.20.0

# --- Logging (optional) ---
tensorboard>=2.16.0
# wandb>=0.16.0        # Uncomment for W&B logging

# --- Utilities ---
tqdm>=4.66.0
